requirements.txt
idotmatrix
requests
Pillow
